moved headerkey class to separate file
