+++ client_secret.json
@@ -1,0 +1,12 @@
+{
+  "type": "service_account",
+  "project_id": "python-gsheet-intergration",
+  "private_key_id": "152d0db16259c15aec24488fba8b221d496f4f8c",
+  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQCN3KjkuQp6FZsN\nbnhKhz/YvkQqXErOlkdGadMKdFO32SiZTzE6O3JWgD20/WaqNvKQwpqNr7PEFV2z\nDrYn9DeVSrAo57JNFKI4wVQTFX35nPvQBaLJP4M9nXg8y28CNK/djHkS7sSGB7iO\n9thq48VCwigUvodsOmmfeFt84iRbwbQNw5fm3PFcoRtb9T8Sq5Pedu7HCmR0vmEO\nPK6wd8OuNZ0NgTsVzjS+0I8zKAsul9LKPJv1U11JlxkCBAOm2KszLw2p767v/qH/\np6c63qI+MLQYbfZ/8+C5OLsASsiL3kDSJ4cXUKPK8AOyb/pGbbdKaQequWtXiMwv\newQUe9aTAgMBAAECggEAO6tq5vnmLD/WPioPMtn8wwAUz/SOCqMvS7DLcgRICnav\nGv0yaZoDl9rNGcnUWxhNoxIXzCzxYjLSfLRdbiTNFGCcBDigyMc7DMBUysopeMrN\nKEyWkivC2jHYF0pCYR0kM/4YFxyTUQUCXi/6C8vOdywPi9owQBe5wbfwU3ETdfw9\nlHm6Sz5bCwsjGmnCkvIwOzxiECKtAQZBvZcuNiqJOpEq2yd05lE8/XBaojM478++\nqp55xX3Cq1M6yo1iLmzISBbhYvyqcsEmpe6Ro06wMlCEfGdCiVSRZ5H8OfUv90Mk\nvTLG1w9i0aOf6zqIt+sWT4XB+1VkV7MIBGxHmNb8cQKBgQDCfqWEGrVbtk+MmT1y\nkGqb7j9ARLQFlnJ3oTmfdB9Cw9M8fTIDVurnAd1kGCUPe9qyo+xT4ru9/Yr2TUsa\n9KD87zj+ei3GKm4JUNqLm08eh9g6EfKtC957BY64VUUfmv+2HbkMIYoaevwpOMq/\nouStzucELy/caE8Qq9agSTgZXQKBgQC6uR5sT21FI4DFOPyWGaPhbrd0vYsNphTC\nengheyea2kb8irdEP28+XxXAc2LeP9rNQJWrPMb7w9RbotWUbfNXhMWy0j0mf0k5\nPdgDvSBy7lahZI8rp+mJxT2K3mu0Or6rZCRyEJf99q8GE/zln4s1QTknaZBb+hjN\n2fgxmQSArwKBgQCmLspD5wJA+7RawPC0+9pro5Ud19Yt1A+4UeTHbtPrGyFZVOuv\n2MtbWMeO5+Z6pWwHKXMbdebOfl4G96INxSntFIYRQdnssYJVqWUi1jAzOQUzEPTW\nkQam/Oembw1h5PFK7XiYeevjlmPw4Nq0BX35WIkhlSQC4ol0CC+ZwxYQ7QKBgA9Q\nZXVeokicW9BRFy48VhuEGBnD/eaVq3gz0Sx1zE6//jhyLAu15CrKY59QF1jBBfH1\njRDpQCyf6AfH+BpfVijkY86BvbAtmlTOhKvx0/p5H7PG6pMbB2VcPurnyQ+UDdVC\ntlaAk+apV7t44jH2/tTUT5qy5rsUbPpdGSUANsVTAoGAHAIGGswWDttI2iGKTe9O\nf2xCmxFBXlqjC2bXE92e7nVrWnWn0sENCMnOtWQssE4gR6UVemvfSYjoO7vUheKP\noByQv1MPR6mFcGthdOfmObwRCaLl85wFg1Ocyhq3dubdDjQembV2Byk35BFwtC+S\nXRD/i7yKf/8Rz8VrsZIkiKQ=\n-----END PRIVATE KEY-----\n",
+  "client_email": "pythondrive@python-gsheet-intergration.iam.gserviceaccount.com",
+  "client_id": "105366672871066320609",
+  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+  "token_uri": "https://accounts.google.com/o/oauth2/token",
+  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/pythondrive%40python-gsheet-intergration.iam.gserviceaccount.com"
+}
